% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferox}
\alias{ferox}
\title{Loch Rannoch Ferox Trout Mark-Recapture Data}
\format{A data frame with the following columns:
\describe{
  \item{Date}{The date of capture or recapture as a Date.}
  \item{Fish}{The individual fish identifier as a factor.}
  \item{Length}{The fork length of the fish at (re)capture (in mm) as an integer.}
  \item{Mass}{The wet mass of the fish at (re)capture (in kg) as a numeric.}
  \item{Age}{The scale age of the fish at (re)capture (in yr) as an integer.}
  \item{Easting}{The easting (in km of projection EPSG:27700) as a numeric.}
  \item{Northing}{The northing (in km of projection EPSG:27700) as a numeric.}
}}
\usage{
ferox
}
\description{
Loch Rannoch Ferox Trout Mark-Recapture Data
}
\seealso{
\code{\link{ranmrdata}}
}
\keyword{datasets}

