% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scotland}
\alias{plot_scotland}
\title{Plot Scotland}
\usage{
plot_scotland()
}
\value{
A ggplot2 object.
}
\description{
Plots Scotland and Western Isles with Loch Rannoch shaded.
}
\examples{
plot_scotland()
}

