% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{power_station}
\alias{power_station}
\title{Rannoch Power Station}
\format{A SpatialPoints object of the location of Rannoch Power Station.
The projection is EPSG:27700.}
\usage{
power_station
}
\description{
Rannoch Power Station
}
\seealso{
\code{\link{ranmrdata}}
}
\keyword{datasets}

