% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ranmrdata}
\alias{ranmrdata}
\alias{ranmrdata-package}
\title{Loch Rannoch Ferox Trout Mark Recapture Data}
\description{
\code{ranmrdata} is the companion R data package for Thorne et al.'s
manuscript The Low Abundance and High Catchability of Large Piscivorous Ferox
Trout (Salmo trutta) in Loch Rannoch, Scotland.
}
\details{
For a list of the exported functions and data objects type:
 \code{objects("package:ranmrdata")}. For more information on an
 object type: \code{?object_name}.
For a description of the data objects type:
 \code{data(package = "ranmrdata")}.
}

