% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland}
\alias{scotland}
\title{Scotland and Western Isles European Electoral Borders}
\format{A SpatialPolygons object of Scotland and the Western Isles
based on the European Electoral Borders.
The projection is EPSG:27700.}
\source{
\url{https://certificates.theodi.org/datasets/620/certificate}
}
\usage{
scotland
}
\description{
Scotland and Western Isles European Electoral Borders
}
\seealso{
\code{\link{ranmrdata}}
}
\keyword{datasets}

