% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferox}
\alias{ferox}
\title{Loch Rannoch Ferox Trout Mark-Recapture Data}
\format{
A tbl data frame with 84 rows and 7 columns:
\describe{
  \item{Date}{The date of capture or recapture (date).}
  \item{Fish}{The individual fish identifier (fctr).}
  \item{Length}{The fork length of the fish at (re)capture (in mm) (int).}
  \item{Mass}{The wet mass of the fish at (re)capture (in kg) (dbl).}
  \item{Age}{The scale age of the fish at (re)capture (in yr) (int).}
  \item{Latitude}{The latitude (WGS84) (dbl).}
  \item{Longitude}{The longitude (WGS84) (dbl).}
}
}
\usage{
ferox
}
\description{
Loch Rannoch Ferox Trout Mark-Recapture Data
}
\keyword{datasets}
