% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rannoch}
\alias{rannoch}
\title{Loch Rannoch Shoreline}
\format{A SpatialPolygons object of the shoreline of Loch Rannoch, Scotland.
The projection is EPSG:32630.}
\usage{
rannoch
}
\description{
Loch Rannoch Shoreline
}
\seealso{
\code{\link{ranmrdata}}
}
\keyword{datasets}

